+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Better error message for missing serializer.
 - Fix generation when there are nested multi-parameter generics.
 - Use cascades in generated code as suggested by lint.
+- Allow users to define any factory that references the generated
+  implementation.
+- Add example of a simpler factory for a one-field class.
 
 ## 0.5.4